Add code for save/load, predicting and metrics
New modules include functions and types for saving forest models into a .rfc file as well as loading the file as a tree into memory again as a forest type. Trees are stored into the file sequentially in postorder. Also added modules for predicting, they contain the necessary functions to perform predictions with a trained model. Finally, another module was added to test the model and print the metrics such as accuracy, precision, recall and F1 score.