Update random forest with train test split function
Random forest train code now contains a function for splitting the dataset into a training dataset and a testing dataset. The user can define which percentage of the dataset will be for testing (the other portion will be for training) and if the rows of the dataset should be shuffled.